+++ app/Http/Controllers/ApiController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\MenuItem;
+use App\Models\MenuItem;
 use App\Repositories\UserRepository;
 use Illuminate\Http\Request;
