+++ app/Models/MenuItem.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
@@ -31,12 +31,12 @@ class MenuItem extends Model
 
     public function partner()
     {
-        return $this->hasOne('App\MenuItem', 'id', 'parent_id');
+        return $this->hasOne('App\Models\MenuItem', 'id', 'parent_id');
     }
 
     public function afterItem()
     {
-        return $this->hasOne('App\MenuItem', 'id', 'after');
+        return $this->hasOne('App\Models\MenuItem', 'id', 'after');
     }
 
     //return the menu items wich aren't sub menus
