+++ database/seeders/ClearSeeder.php
@@ -3,13 +3,13 @@
 namespace Database\Seeders;
 
 use App\CustomClasses\MailList\MailListFacade;
-use App\MenuItem;
 use App\Models\AgendaItem;
 use App\Models\AgendaItemCategory;
 use App\Models\ApplicationForm\ApplicationForm;
 use App\Models\ApplicationForm\ApplicationFormRow;
 use App\Models\Book;
 use App\Models\Certificate;
+use App\Models\MenuItem;
 use App\Models\UserEventLogEntry;
 use App\NewsItem;
 use App\Rol;
